+++ emp-tool/io/tls_io_channel.h
@@ -294,11 +294,11 @@ class TLSIO : public IOChannel { public:
 		}
 	}
 
-	void send_data_internal(const void *data, size_t len) override {
+	void send_data_internal(const void *data, int64_t len) override {
 #ifndef NDEBUG
 		touch_guard _g(_in_use, "send_data");
 #endif
-		if (len + send_ptr <= (size_t)NETWORK_STAGING_BUFFER_SIZE) {
+		if (len + (int64_t)send_ptr <= (int64_t)NETWORK_STAGING_BUFFER_SIZE) {
 			memcpy(send_buf + send_ptr, data, len);
 			send_ptr += len;
 		} else {
@@ -308,7 +308,7 @@ class TLSIO : public IOChannel { public:
 		send_dirty = true;
 	}
 
-	void recv_data_internal(void *data, size_t len) override {
+	void recv_data_internal(void *data, int64_t len) override {
 #ifndef NDEBUG
 		touch_guard _g(_in_use, "recv_data");
 #endif
@@ -318,7 +318,7 @@ class TLSIO : public IOChannel { public:
 		// drain, so this has to be explicit.
 		flush_unlocked();
 		bytes_recv += len;
-		size_t got = 0;
+		int64_t got = 0;
 		while (got < len) {
 			if (recv_ptr == recv_fill) {
 				size_t n = ssl_read_some(recv_buf, NETWORK_STAGING_BUFFER_SIZE);
@@ -326,7 +326,7 @@ class TLSIO : public IOChannel { public:
 				recv_ptr = 0;
 				recv_fill = n;
 			}
-			size_t take = recv_fill - recv_ptr;
+			int64_t take = recv_fill - recv_ptr;
 			if (take > len - got) take = len - got;
 			memcpy((char *)data + got, recv_buf + recv_ptr, take);
 			recv_ptr += take;
